Revert from uint64_t to int, for MATIO compatibility

+++ warp_aux.h
@@ -17,15 +17,15 @@ do { \
     } \
 } while (0)
 
-__global__ void label_propagation_kernel(uint64_t* labels, int* active, int* next_active, 
-                                        const uint64_t* indices, const uint64_t* ind_ptr, 
-                                        uint64_t n_nodes, int* changed_flag);
+__global__ void label_propagation_kernel(int* labels, int* active, int* next_active, 
+                                        const int* indices, const int* ind_ptr, 
+                                        int n_nodes, int* changed_flag);
 
-__global__ void initialize_kernel (uint64_t* labels, int* active, int *next_active, uint64_t n_nodes);
+__global__ void initialize_kernel (int* labels, int* active, int *next_active, int n_nodes);
 
-__global__ void reset_active_kernel(int* next_active, uint64_t n_nodes);
+__global__ void reset_active_kernel(int* next_active, int n_nodes);
 
-void open_matrix (char* name, uint64_t** indices, uint64_t** ind_ptr, uint64_t* n_nodes,
+void open_matrix (char* name, int** indices, int** ind_ptr, int* n_nodes,
                 mat_t** matfp_out, matvar_t** problem_out);
 
 #endif
